refactor(groups): move groups-discovery service into discovery/ subdirectory
Consistent with the slice-per-subdirectory layout applied to the rest of the
groups module. Schema and dto imports updated to @/ aliases.
